+++ actix-protobuf/CHANGES.md
@@ -1,10 +1,10 @@
 # Changes
 
-## unreleased (for alpha version)
+## 0.6.0-alpha.1 (2020-07-06)
+
+* Update `actix-web` to 3.0.0-alpha.3
 
 * Minimum supported Rust version(MSRV) is now 1.40.0.
-
-## unreleased
 
 * Minimize `futures` dependency